- Resolve merge conflict in docs/plans/PDSL_AUDIO_DSP.md: take master version,
  which reflects the actual PDSL audio DSP implementation (not the old exploratory
  design that was on HEAD). The master plan covers all landed primitives, multi-channel
  constructs, producer-valued arguments, the MixdownManagerPdslAdapter, and what is
  still deferred (delay_network, production cutover, EfxManager migration).

- Substantially rewrite docs/plans/AUDIO_SCENE_REDESIGN.md to reflect the corrected
  scope and current state of the codebase:

  1. Corrected scope: pattern generation (~278ms of 311ms total per tick) dominates
     render cost; DSP mixdown (~33ms) is secondary. The original plan focused entirely
     on the DSP half. The revised plan covers both.

  2. PDSL is the substrate: not one option among several. The MixdownManagerPdslAdapter,
     efx_channel.pdsl, mixdown_manager.pdsl, and the full set of audio primitives are
     already on master. The original plan's Options A-D (per-channel CellList splitting,
     graph-analysis copy-chain elimination, stream-oriented processing) are moot —
     per-channel kernels fall out naturally from PDSL compilation.

  3. What remains: delay_network primitive (prerequisite for full cutover), production
     cutover of MixdownManager.createCells(), EfxManager migration, variable channel
     count support, and CellList retirement.

  4. Phasing: Phase 1 (DSP cutover, mostly mechanical — substrate exists), Phase 2
     (pattern profiling + targeted fixes, evaluative — profile first then fix), Phase 3
     (pattern restructuring if Phase 2 is insufficient, conditional), Phase 4
     (everything is PDSL, aspirational trajectory).

  5. Long-term direction stated honestly: AudioScene as a single PDSL computation graph,
     without timeline commitments. Intermediate phases preserve this option.

+++ compute/algebra/src/main/java/org/almostrealism/collect/CollectionFeatures.java
@@ -1054,7 +1054,7 @@ public interface CollectionFeatures extends GradientFeatures, CollectionCreation
 											Producer<PackedCollection> collection,
 											Producer<PackedCollection> index) {
 		DefaultTraversableExpressionComputation exp = new DefaultTraversableExpressionComputation("valueAtIndex", shape,
-				args -> CollectionExpression.create(shape, idx -> args[1].getValueAt(args[2].getValueAt(idx))),
+				args -> CollectionExpression.create(shape, idx -> args[1].getValueAt(args[2].getValueAt(idx).toInt())),
 				collection, index);
 		if (shape.getTotalSize() == 1 && Countable.isFixedCount(index)) {
 			exp.setShortCircuit(args -> {
